feat(HTMLImage): add new component
This component is now used instead of plain <Image> as the default
renderer for <img> tags.
Instead of settings the image’s size depending on the style attribute
in HTML, it fetches its dimensions before displaying it.

If its width exceeds the new optionnal imagesMaxWidth prop, images are
resized (and keep the same ratio).